+++ java/src/simoil/estrategias/reinyeccion/EstrategiaReinyeccionNoReinyectar.java
@@ -4,7 +4,7 @@ import simoil.EmprendimientoPetrolifero;
 
 public class EstrategiaReinyeccionNoReinyectar extends EstrategiaReinyeccion{
     @Override
-    public float calcularTotalLitrosReinyeccion(EmprendimientoPetrolifero emprendimientoPetrolifero, float volumenMaximoReinyeccionEnUnDia) {
+    public double calcularTotalLitrosReinyeccion(EmprendimientoPetrolifero emprendimientoPetrolifero, double volumenMaximoReinyeccionEnUnDia) {
         return 0;
     }
 }
